Separated WinMLRunner into 2 separate projects. One builds DLL and one builds EXE. EXE only contains main.cpp
