+++ Tools/WinMLRunner/Run.h
@@ -1,0 +1,3 @@
+#include "CommandLineArgs.h"
+
+__declspec(dllexport) int run(CommandLineArgs& args);
